+++ plugins/commands/publish.go
@@ -5,8 +5,8 @@ import (
 	"github.com/codegangsta/cli"
 	"github.com/jfrog/gofrog/io"
 	"github.com/jfrog/jfrog-cli-core/artifactory/commands/generic"
-	"github.com/jfrog/jfrog-cli-core/artifactory/spec"
 	rtutils "github.com/jfrog/jfrog-cli-core/artifactory/utils"
+	"github.com/jfrog/jfrog-cli-core/common/spec"
 	"github.com/jfrog/jfrog-cli-core/utils/config"
 	"github.com/jfrog/jfrog-cli/plugins/commands/utils"
 	pluginsutils "github.com/jfrog/jfrog-cli/plugins/utils"
